feat: LIBR-544 - rename utils package; update .goreleaser.yaml
